Move off of react router onEnter.
onEnter is deprecated, instead you should use route component lifecycles.
