+++ src/Pages/Init.tsx
@@ -1,9 +1,26 @@
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 
 import { withRouter } from "react-router-dom";
 import Services from "../services";
 
+function redirectTo(to: string) {
+  function replacePath(
+    nextState: { location: { pathname: string } },
+    replace: (replacement: {}) => void
+  ) {
+    replace({
+      pathname: to,
+      state: { nextPathname: nextState.location.pathname },
+    });
+  }
+  return replacePath;
+}
+
 const Init: React.FC = () => {
+  useEffect(() => {
+    redirectTo("/init");
+  }, []);
+
   const [state, setState] = useState({
     uid: "",
     contestants: "",
